+++ src/egon/data/airflow/dags/pipeline.py
@@ -29,6 +29,7 @@ from egon.data.datasets.electricity_demand_timeseries import (
 from egon.data.datasets.era5 import WeatherData
 from egon.data.datasets.etrago_setup import EtragoSetup
 from egon.data.datasets.fill_etrago_gen import Egon_etrago_gen
+from egon.data.datasets.gas_aggregation import GasAggregation
 from egon.data.datasets.gas_areas import GasAreaseGon2035, GasAreaseGon100RE
 from egon.data.datasets.gas_grid import GasNodesandPipes
 from egon.data.datasets.gas_prod import CH4Production
@@ -463,6 +464,13 @@ with airflow.DAG(
     # Assign industrial gas demand eGon100RE
     assign_industrial_gas_demand = IndustrialGasDemandeGon100RE(
         dependencies=[industrial_gas_demand, create_gas_polygons_egon100RE]
+    )
+    # Aggregate gas loads, stores and generators
+    aggrgate_gas = GasAggregation(
+        dependencies=[
+            gas_production_insert_data,
+            insert_data_ch4_storages,
+        ]
     )
 
     # CHP locations